fixed failing test in FIFO read producer
